[API] Rename BatchTreatment to batch-processing
